backend/app/requirements.txt
fastapi
uvicorn
chromadb
langchain-groq
langchain-community
tavily-python
pandas
python-dotenv
sqlalchemy
trafilatura          
google-api-python-client 